+++ examples/supertrader/src/products/SuperTrader/_lib/react/web/UCOutputFieldValue.tsx
@@ -3,7 +3,7 @@ import type { UCOutputFieldValueFragmentProps } from 'libmodulor/react';
 import { UCOutputFieldValue as UCOutputFieldValueBase } from 'libmodulor/react-web-pure';
 import React, { type ReactElement } from 'react';
 
-import { TISIN, TOrderStatus } from '../../../../apps/Trading/index.js';
+import { TISIN, TOrderStatus } from '../../../../../apps/Trading/index.js';
 import Badge from './Badge.js';
 
 export default function UCOutputFieldValue<