+++ examples/supertrader/src/products/SuperTrader/server-nextjs/src/app/page.tsx
@@ -1,0 +1,7 @@
+import React, { type ReactElement } from 'react';
+
+import App from '../../../_lib/react/web/App.js';
+
+export default function Page(): ReactElement {
+    return <App />;
+}